+++ mod/print.lua
@@ -1,4 +1,5 @@
 local __DebugAdapter = __DebugAdapter
+local debug = debug
 local variables = require("__debugadapter__/variables.lua") -- uses pcall
 local normalizeLuaSource = require("__debugadapter__/normalizeLuaSource.lua")
 require("__debugadapter__/evaluate.lua") -- uses pcall